backend/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
beautifulsoup4==4.14.2
cachetools==5.5.2
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
fastapi==0.115.0
google-ai-generativelanguage==0.6.6
google-api-core==2.26.0
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
google-generativeai==0.7.2
googleapis-common-protos==1.70.0
grpcio==1.75.1
grpcio-status==1.62.3
h11==0.16.0
httplib2==0.31.0
httptools==0.7.1
idna==3.11
lxml==6.0.2
numpy==2.3.4
oauthlib==3.3.1
pillow==11.2.1
proto-plus==1.26.1
protobuf==4.25.8
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.2
pydantic_core==2.41.4
pyparsing==3.2.5
python-dotenv==1.0.0
PyYAML==6.0.3
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
sniffio==1.3.1
soupsieve==2.8
starlette==0.38.6
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.30.6
watchfiles==1.1.1
websockets==15.0.1
